// infrastructure/client-library/src/types.ts
export interface RpcProvider {
  baseUrl: string;
  getTxById(txId: string): Promise<EsploraTransaction>;
  getAddressTxs(address: string): Promise<EsploraTransaction[]>;
  getTxOutput(txId: string, index: number): Promise<OrdOutput>;
  getInscriptionById(inscriptionId: string): Promise<OrdInscription>;
  getRuneById(runeId: string): Promise<OrdRune>;
}

export interface EsploraTransaction {
  txid: string;
  version: number;
  locktime: number;
  vin: {
    txid: string;
    vout: number;
    prevout: {
      scriptpubkey: string;
      scriptpubkey_asm: string;
      scriptpubkey_type: string;
      scriptpubkey_address: string;
      value: number;
    };
    scriptsig: string;
    scriptsig_asm: string;
    witness: string[];
    is_coinbase: boolean;
    sequence: number;
  }[];
  vout: {
    scriptpubkey: string;
    scriptpubkey_asm: string;
    scriptpubkey_type: string;
    scriptpubkey_address: string;
    value: number;
  }[];
  size: number;
  weight: number;
  fee: number;
  status: {
    confirmed: boolean;
    block_height?: number;
    block_hash?: string;
    block_time?: number;
  };
}

export type OrdOutputRune = {
  amount: number;
  divisibility: number;
};

export interface OrdOutput {
  address: string;
  indexed: boolean;
  inscriptions: string[];
  runes: Record<string, OrdOutputRune>;
  sat_ranges: number[][];
  script_pubkey: string;
  spent: boolean;
  transaction: string;
  value: number;
}

export interface OrdRune {
  entry: {
    block: number;
    burned: number;
    divisibility: number;
    etching: string;
    mints: number;
    number: number;
    premine: number;
    spaced_rune: string;
    symbol: string;
    terms: {
      amount: number;
      cap: number;
      height: any[];
      offset: any[];
    };
    timestamp: number;
    turbo: boolean;
  };
  id: string;
  mintable: boolean;
  parent: string;
}

export type OrdInscription = {
  id: string;
  content_type: string;
  content: string;
};
// ...
export interface BackendServiceConfig {
  url: string;
  wsUrl?: string;
}